+++ bika/lims/jsonapi/v1/__init__.py
