+++ bika/lims/jsonapi/v1/configure.zcml
@@ -8,7 +8,7 @@
       description="Better jsonapi error handler."
       module="plone.jsonapi.core.browser.decorators"
       original="handle_errors"
-      replacement="bika.lims.jsonapi.handle_errors"
+      replacement="bika.lims.jsonapi.v1.handle_errors"
       />
 
   <utility
