+++ bika/lims/jsonapi/v1/update.py
@@ -3,7 +3,7 @@
 # Copyright 2011-2016 by it's authors.
 # Some rights reserved. See LICENSE.txt, AUTHORS.txt.
 
-from bika.lims.jsonapi import set_fields_from_request
+from bika.lims.jsonapi.v1 import set_fields_from_request
 from Products.CMFCore.utils import getToolByName
 from plone.jsonapi.core import router
 from plone.jsonapi.core.interfaces import IRouteProvider
