+++ bika/lims/workflow.py
@@ -7,7 +7,7 @@ from bika.lims import enum
 from bika.lims import PMF
 from bika.lims.browser import ulocalized_time
 from bika.lims.interfaces import IJSONReadExtender
-from bika.lims.jsonapi import get_include_fields
+from bika.lims.jsonapi.v1 import get_include_fields
 from bika.lims.utils import changeWorkflowState
 from bika.lims.utils import t
 from bika.lims import logger
